4.21

+++ part4/blog_app/controllers/blogs.js
@@ -49,7 +49,16 @@ blogRouter.post('/', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
-  const blog = await Blog.findByIdAndRemove(request.params.id)
+  const blogToDelete = await Blog.findById(request.params.id)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+  if (blogToDelete.user.toString() !== user.id.toString()) {
+    return response.status(401).json({ error: 'unauthorized user' })
+  }
+  await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
